pyproject.toml
[project]
name = "antifraud2gis"
version = "0.1.0"
dependencies = [
    "requests",
    "loguru",
    "rich",
    "argalias",
    "python-dotenv"
    ]
requires-python = ">=3.8"
authors = [{ name = "Yaroslav Polyakov", email = "yaroslaff@gmail.com" }]

[project.urls]
Homepage = "https://github.com/yaroslaff/scrape2gis"
Repository = "https://github.com/yaroslaff/scrape2gis"


[project.scripts]
af2gis = "antifraud2gis.main:main"

[build-system]
# requires = ["setuptools"]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
